#1052 Move property value change notifications wrappers (for child objects & collections) from ModelBase to ChildAwareModelBase
