+++ src/Catel.Core/Catel.Core.Shared/Catel.Core.Shared.projitems
@@ -73,6 +73,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)CoreModule.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\Attributes\ExcludeFromValidationAttribute.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\Attributes\ValidateModelAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\ChildAwareModelBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\Extensions\IModelExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\Extensions\IModelExtensions.serialization.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\Extensions\ISavableModelExtensions.cs" />
