+++ src/Catel.Core/Catel.Core.Shared/Data/ChildAwareModelBase.cs
@@ -1,0 +1,256 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="ChildAwareModelBase.cs" company="Catel development team">
+//   Copyright (c) 2008 - 2017 Catel development team. All rights reserved.
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+
+namespace Catel.Data
+{
+    using System;
+    using System.Collections;
+    using System.Collections.Generic;
+    using System.Collections.Specialized;
+    using System.ComponentModel;
+
+#if NET
+    using System.Runtime.Serialization;
+#endif
+
+    /// <summary>
+    /// Class that is aware of changes of child objects by using the <see cref="ChangeNotificationWrapper"/>.
+    /// </summary>
+    public abstract class ChildAwareModelBase : ModelBase
+    {
+        /// <summary>
+        /// The change notification wrappers for all property values.
+        /// </summary>
+#if NET
+        [field: NonSerialized]
+#endif
+        private Dictionary<string, ChangeNotificationWrapper> _propertyValueChangeNotificationWrappers;
+
+        /// <summary>
+        /// Initializes the <see cref="ChildAwareModelBase"/> class.
+        /// </summary>
+        static ChildAwareModelBase()
+        {
+            DefaultDisableEventSubscriptionsOfChildValuesValue = false;
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ChildAwareModelBase"/> class.
+        /// </summary>
+        protected ChildAwareModelBase()
+        {
+            InitializeChildAwareModelBase();
+        }
+
+#if NET
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ModelBase"/> class.
+        /// <para />
+        /// Only constructor for the ModelBase.
+        /// </summary>
+        /// <param name="info">SerializationInfo object, null if this is the first time construction.</param>
+        /// <param name="context">StreamingContext object, simple pass a default new StreamingContext() if this is the first time construction.</param>
+        /// <remarks>
+        /// Call this method, even when constructing the object for the first time (thus not deserializing).
+        /// </remarks>
+        protected ChildAwareModelBase(SerializationInfo info, StreamingContext context)
+            : base(info, context)
+        {
+            InitializeChildAwareModelBase();
+        }
+#endif
+
+        #region Properties
+        /// <summary>
+        /// Gets or sets a value indicating whether event subscriptions of child values should be disabled.
+        /// <para />
+        /// The default value is <c>false</c>.
+        /// </summary>
+        /// <value><c>true</c> if event subscriptions of child values should be disabled; otherwise, <c>false</c>.</value>
+        protected bool DisableEventSubscriptionsOfChildValues { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether event subscriptions of child values should be disabled.
+        /// </summary>
+        public static bool DefaultDisableEventSubscriptionsOfChildValuesValue { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether this object should handle (thus invoke the specific events) when
+        /// a property of collection value has changed.
+        /// </summary>
+#if NET
+        [Browsable(false)]
+#endif
+        protected bool HandlePropertyAndCollectionChanges { get; set; }
+
+        #endregion
+
+        /// <summary>
+        /// Initializes this instance.
+        /// </summary>
+        private void InitializeChildAwareModelBase()
+        {
+            HandlePropertyAndCollectionChanges = true;
+            DisableEventSubscriptionsOfChildValues = DefaultDisableEventSubscriptionsOfChildValuesValue;
+        }
+
+        /// <summary>
+        /// Raises the <see cref="E:PropertyChanged" /> event.
+        /// </summary>
+        /// <param name="e">The <see cref="AdvancedPropertyChangedEventArgs"/> instance containing the event data.</param>
+        protected override void OnPropertyChanged(AdvancedPropertyChangedEventArgs e)
+        {
+            base.OnPropertyChanged(e);
+
+            if (HandlePropertyAndCollectionChanges)
+            {
+                var sender = e.OriginalSender;
+
+                if (ReferenceEquals(this, sender))
+                {
+                    //if (!isRefreshCallOnly)
+                    //{
+                        SuspensionContext callbackSuspensionContext;
+
+                        lock (_lock)
+                        {
+                            callbackSuspensionContext = _changeCallbacksSuspensionContext;
+                        }
+
+                        if (callbackSuspensionContext != null)
+                        {
+                            callbackSuspensionContext.Add(e.PropertyName);
+                        }
+                        else if (IsPropertyRegistered(e.PropertyName))
+                        {
+                            var propertyData = GetPropertyData(e.PropertyName);
+
+                            var handler = propertyData.PropertyChangedEventHandler;
+                            if (handler != null)
+                            {
+                                handler(this, e);
+                            }
+                        }
+                    //}
+
+                    if (!DisablePropertyChangeNotifications)
+                    {
+                        // Explicitly call base because we have overridden the behavior
+                        base.RaisePropertyChanged(this, e);
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Sets the value fast without checking for any constraints or additional logic such as change notifications. This 
+        /// means that if this method is used incorrectly, it can throw random exceptions.
+        /// <para />
+        /// This is a wrapper around the _propertyValues field. Don't use the field directly, always use
+        /// this method because it takes care of locking and event subscriptions.
+        /// </summary>
+        /// <param name="propertyName">Name of the property.</param>
+        /// <param name="value">The value.</param>
+        protected override void SetValueToPropertyBag(string propertyName, object value)
+        {
+            base.SetValueToPropertyBag(propertyName, value);
+
+            HandleObjectEventsSubscription(propertyName, value);
+        }
+
+        /// <summary>
+        /// Handles the object events subscription. This means that the old value will be removed from the event subscriptions, and
+        /// the new value will be subscribed to.
+        /// </summary>
+        /// <param name="propertyName">Name of the property.</param>
+        /// <param name="propertyValue">The property value.</param>
+        private void HandleObjectEventsSubscription(string propertyName, object propertyValue)
+        {
+            if (DisableEventSubscriptionsOfChildValues)
+            {
+                return;
+            }
+
+            lock (_lock)
+            {
+                ChangeNotificationWrapper oldWrapper;
+
+                if (_propertyValueChangeNotificationWrappers == null)
+                {
+                    _propertyValueChangeNotificationWrappers = new Dictionary<string, ChangeNotificationWrapper>();
+                }
+
+                if (_propertyValueChangeNotificationWrappers.TryGetValue(propertyName, out oldWrapper))
+                {
+                    oldWrapper.PropertyChanged -= OnPropertyObjectPropertyChanged;
+                    oldWrapper.CollectionChanged -= OnPropertyObjectCollectionChanged;
+                    oldWrapper.CollectionItemPropertyChanged -= OnPropertyObjectCollectionItemPropertyChanged;
+                    oldWrapper.UnsubscribeFromAllEvents();
+                }
+
+                if (!ChangeNotificationWrapper.IsUsefulForObject(propertyValue))
+                {
+                    if (oldWrapper != null)
+                    {
+                        _propertyValueChangeNotificationWrappers.Remove(propertyName);
+                    }
+                }
+                else
+                {
+                    var wrapper = new ChangeNotificationWrapper(propertyValue);
+                    wrapper.PropertyChanged += OnPropertyObjectPropertyChanged;
+                    wrapper.CollectionChanged += OnPropertyObjectCollectionChanged;
+                    wrapper.CollectionItemPropertyChanged += OnPropertyObjectCollectionItemPropertyChanged;
+                    _propertyValueChangeNotificationWrappers[propertyName] = wrapper;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Called when a property that implements <see cref="INotifyPropertyChanged"/> raises the event.
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="e">The <see cref="PropertyChangedEventArgs"/> instance containing the event data.</param>
+        protected virtual void OnPropertyObjectPropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            // It is possible that the sender used string.Empty or null for the property name, then exit
+            if (string.IsNullOrEmpty(e.PropertyName))
+            {
+                return;
+            }
+
+            RaisePropertyChanged(sender, e, true, false);
+        }
+
+        /// <summary>
+        /// Called when a property that implements <see cref="INotifyCollectionChanged"/> raises the event.
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="e">The <see cref="NotifyCollectionChangedEventArgs"/> instance containing the event data.</param>
+        protected virtual void OnPropertyObjectCollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
+        {
+            SetDirty(string.Empty);
+        }
+
+        /// <summary>
+        /// Called when a property inside a collection that implements <see cref="INotifyCollectionChanged"/> that implements
+        /// <see cref="INotifyPropertyChanged"/> raises the event.
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="e">The <see cref="PropertyChangedEventArgs"/> instance containing the event data.</param>
+        protected virtual void OnPropertyObjectCollectionItemPropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            if (string.Equals(e.PropertyName, "IsDirty", StringComparison.Ordinal))
+            {
+                return;
+            }
+
+            SetDirty(string.Empty);
+        }
+
+    }
+}
