+++ src/Catel.Test/Catel.Test.Shared/Catel.Test.Shared.projitems
@@ -32,6 +32,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Data\BoxingCacheFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\CatelTypeInfoFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ChangeNotificationWrapperFacts.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\ChildAwareModelBaseFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\CompositeValidatorFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\CompositeValidatorProviderFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\EventArgs\AdvancedPropertyChangedEventArgsFacts.cs" />
