+++ src/Catel.Test/Catel.Test.Shared/Data/ChildAwareModelBaseFacts.cs
@@ -1,0 +1,101 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="ChildAwareModelBaseFacts.cs" company="Catel development team">
+//   Copyright (c) 2008 - 2017 Catel development team. All rights reserved.
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+
+namespace Catel.Test.Data
+{
+    using System.Collections.Generic;
+    using System.Collections.ObjectModel;
+    using System.ComponentModel;
+    using System.Windows.Data;
+    using Catel.Data;
+    using Catel.MVVM;
+    using NUnit.Framework;
+
+    [TestFixture]
+    public class ChildAwareModelBaseFacts
+    {
+        public class ModelWithObservableCollection : ChildAwareModelBase
+        {
+            /// <summary>
+            /// Gets or sets the property value.
+            /// </summary>
+            public ObservableCollection<int> Collection
+            {
+                get { return GetValue<ObservableCollection<int>>(CollectionProperty); }
+                set { SetValue(CollectionProperty, value); }
+            }
+
+            /// <summary>
+            /// Register the Collection property so it is known in the class.
+            /// </summary>
+            public static readonly PropertyData CollectionProperty = RegisterProperty("Collection", typeof(ObservableCollection<int>), () => new ObservableCollection<int>());
+
+            public bool HasCollectionChanged { get; private set; }
+
+            public bool HasPropertyChanged { get; private set; }
+
+            protected override void OnPropertyObjectCollectionChanged(object sender, System.Collections.Specialized.NotifyCollectionChangedEventArgs e)
+            {
+                HasCollectionChanged = true;
+            }
+
+            protected override void OnPropertyObjectPropertyChanged(object sender, PropertyChangedEventArgs e)
+            {
+                HasPropertyChanged = true;
+            }
+        }
+
+#if NET
+        public class ModelWithCollectionViewSource : ViewModelBase
+        {
+            /// <summary>
+            /// Gets or sets the property value.
+            /// </summary>
+            public CollectionView Collection
+            {
+                get { return GetValue<CollectionView>(CollectionProperty); }
+                set { SetValue(CollectionProperty, value); }
+            }
+
+            /// <summary>
+            /// Register the Collection property so it is known in the class.
+            /// </summary>
+            public static readonly PropertyData CollectionProperty = RegisterProperty("Collection", typeof(CollectionView), () => new CollectionView(new List<int>() { 1, 2, 3 }));
+        }
+#endif
+
+        [TestCase]
+        public void AllowsRegistrationOfObservableCollection()
+        {
+            var model = new ModelWithObservableCollection();
+            model.Collection = new ObservableCollection<int>(new List<int>() { 1, 2, 3 });
+
+            model.Collection.Add(4);
+
+            Assert.IsTrue(model.HasCollectionChanged);
+        }
+
+#if NET
+        [TestCase]
+        public void AllowsRegistrationOfCollectionViewSource()
+        {
+            var model = new ModelWithCollectionViewSource();
+            model.Collection = new CollectionView(new List<int>() { 1, 2, 3 });
+        }
+#endif
+
+        [TestCase]
+        public void RegistersChangeNotificationsOfDefaultValues()
+        {
+            var model = new ModelWithObservableCollection();
+
+            model.Collection.Add(4);
+
+            Assert.IsTrue(model.HasCollectionChanged);
+        }
+    }
+}
