+++ src/Catel.Test/Catel.Test.Shared/Data/ModelBaseFacts.cs
@@ -162,90 +162,6 @@ namespace Catel.Test.Data
 #endif
 
         [TestFixture]
-        public class ThePropertiesWithEventSubscriptionsChecks
-        {
-            public class ModelWithObservableCollection : ModelBase
-            {
-                /// <summary>
-                /// Gets or sets the property value.
-                /// </summary>
-                public ObservableCollection<int> Collection
-                {
-                    get { return GetValue<ObservableCollection<int>>(CollectionProperty); }
-                    set { SetValue(CollectionProperty, value); }
-                }
-
-                /// <summary>
-                /// Register the Collection property so it is known in the class.
-                /// </summary>
-                public static readonly PropertyData CollectionProperty = RegisterProperty("Collection", typeof(ObservableCollection<int>), () => new ObservableCollection<int>());
-
-                public bool HasCollectionChanged { get; private set; }
-
-                public bool HasPropertyChanged { get; private set; }
-
-                protected override void OnPropertyObjectCollectionChanged(object sender, System.Collections.Specialized.NotifyCollectionChangedEventArgs e)
-                {
-                    HasCollectionChanged = true;
-                }
-
-                protected override void OnPropertyObjectPropertyChanged(object sender, PropertyChangedEventArgs e)
-                {
-                    HasPropertyChanged = true;
-                }
-            }
-
-#if NET
-            public class ModelWithCollectionViewSource : ViewModelBase
-            {
-                /// <summary>
-                /// Gets or sets the property value.
-                /// </summary>
-                public CollectionView Collection
-                {
-                    get { return GetValue<CollectionView>(CollectionProperty); }
-                    set { SetValue(CollectionProperty, value); }
-                }
-
-                /// <summary>
-                /// Register the Collection property so it is known in the class.
-                /// </summary>
-                public static readonly PropertyData CollectionProperty = RegisterProperty("Collection", typeof(CollectionView), () => new CollectionView(new List<int>() { 1, 2, 3 }));
-            }
-#endif
-
-            [TestCase]
-            public void AllowsRegistrationOfObservableCollection()
-            {
-                var model = new ModelWithObservableCollection();
-                model.Collection = new ObservableCollection<int>(new List<int>() { 1, 2, 3 });
-
-                model.Collection.Add(4);
-
-                Assert.IsTrue(model.HasCollectionChanged);
-            }
-
-#if NET
-            [TestCase]
-            public void AllowsRegistrationOfCollectionViewSource()
-            {
-                var model = new ModelWithCollectionViewSource();
-                model.Collection = new CollectionView(new List<int>() { 1, 2, 3 });
-            }
-#endif
-
-            [TestCase]
-            public void RegistersChangeNotificationsOfDefaultValues()
-            {
-                var model = new ModelWithObservableCollection();
-
-                model.Collection.Add(4);
-
-                Assert.IsTrue(model.HasCollectionChanged);
-            }
-        }
-
-        [TestFixture]
         public class TheEqualsChecks
         {
             public interface ITestModel
